+++ client/src/CHANGELOG.md
@@ -1,6 +1,17 @@
 # Update History
 
 
+## v1.0.4
+### 2025-12-01
+- Add Difficulty Setting
+- Add Difficulty Base Ruleset
+- Fix Rulesets
+
+## v1.0.3
+### 2025-11-30
+- Fix Indicator
+- Fix Reconnection Logic
+
 ## v1.0.2
 ### 2025-11-30
 - Add Profile Page